pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-nexus"
version = "0.1.0"
description = "Agent Nexus: Multi-model AI swarm orchestrated through Discord"
readme = {text = "", content-type = "text/markdown"}
authors = [{ name = "Anthony Maio" }]
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    # Discord
    "discord.py>=2.3",
    # HTTP client
    "aiohttp>=3.9",
    "httpx>=0.27",
    # Vector DB
    "qdrant-client>=1.8",
    # Cache
    "redis>=5.0",
    # Environment
    "python-dotenv>=1.0",
    # Data validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Continuity Core dependencies
    "neo4j>=5.23",
    "psycopg[binary]>=3.1",
    "requests>=2.32",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
local = [
    # For local embeddings via sentence-transformers
    "sentence-transformers>=2.7",
]

[project.scripts]
nexus-setup = "setup.setup:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["nexus*", "continuity_core*", "synthesis*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
